Split praxis into header and code
